backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
opencv-python>=4.9.0
Pillow>=10.2.0
